Add files via upload
This c++ program defines a chocolate class with three public data members-brand, price, flavour. The class also has a member function displayInfo( ) that prints out the chocolate information .In the main ( ) function two objects A1 and A2 are created from the chocolate class. The program assigns values to the data members of each chocolate (A1 and A2). It then calls the displayInfo( ) function to display the properties of both chocolates .This program demonstrates the basic principles of object oriented programming by creating and using objects from a class in c++